refactored cellForRowAt logic to the ViewModel
